// src/generated/api.ts
export interface components {
    schemas: {
// ...
        Member: {
            /** @description Unique member id */
            _id: components["schemas"]["MemberCompositeKey"];
            /** @description Time at which this user joined the server */
            joined_at: components["schemas"]["ISO8601 Timestamp"];
            /** @description Member's nickname */
            nickname?: string | null;
            /** @description Member's pronouns */
            pronouns?: string | null;
            /** @description Avatar attachment */
            avatar?: components["schemas"]["File"] | null;
            /** @description Member's roles */
            roles?: string[];
            /** @description Timestamp this member is timed out until */
            timeout?: components["schemas"]["ISO8601 Timestamp"] | null;
            /** @description Whether the member is server-wide voice muted */
            can_publish?: boolean;
            /** @description Whether the member is server-wide voice deafened */
            can_receive?: boolean;
        };
// ...
        PartialMember: {
            id?: components["schemas"]["MemberCompositeKey"] | null;
            joined_at?: components["schemas"]["ISO8601 Timestamp"] | null;
            nickname?: string | null;
            pronouns?: string | null;
            avatar?: components["schemas"]["File"] | null;
            roles?: string[] | null;
            timeout?: components["schemas"]["ISO8601 Timestamp"] | null;
            can_publish?: boolean | null;
            can_receive?: boolean | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}